Separate Uno from K2 and ANE support
