+++ apps/omlx-mac/Sources/AppView/ViewModels/ModelSettingsScreenVM.swift
@@ -54,7 +54,7 @@ final class ModelSettingsScreenVM {
         case dflashVerifyMode, dflashDraftWindowSize, dflashDraftSinkSize, dflashBlockSize
         case dflashInMemoryCache, dflashInMemoryCacheGib, dflashInMemoryCacheMaxEntries
         case dflashSsdCache, dflashSsdCacheGib
-        case mtpEnabled, unoEnabled, unoAdapterModel
+        case mtpEnabled
         case vlmMtpEnabled, vlmMtpDraftModel, vlmMtpDraftBlockSize
     }
 
@@ -330,8 +330,6 @@ final class ModelSettingsScreenVM {
     var dflashSsdCacheGib: String = "20"
 
     // Experimental: native MTP
-    var unoEnabled: Bool = false
-    var unoAdapterModel: String = ""
     var mtpEnabled: Bool = false
 
     // Experimental: VLM MTP (assistant-drafter speculative decoding for VLMs).
@@ -397,49 +395,6 @@ final class ModelSettingsScreenVM {
         isK2Base ? ["low", "medium", "high"] : ChatTemplateKwargsCodec.reasoningEffortPresets
     }
 
-    var unoAdapterCandidates: [ModelDTO] {
-        guard isK2Base, let base = model?.unoBaseModelId else { return [] }
-        return allModels.filter {
-            $0.configModelType == "k2_horizon_uno" && $0.unoBaseModelId == base
-        }
-    }
-
-    func unoAdapterModelOptions() -> [(String, String)] {
-        [("", String(localized: "settings.uno.select", defaultValue: "Select a Uno adapter"))]
-            + unoAdapterCandidates.map { ($0.id, $0.displayName ?? $0.id) }
-    }
-
-    var unoConflictReason: String? {
-        let penalties = [(minP, 0.0), (repetitionPenalty, 1.0), (presencePenalty, 0.0)]
-        if mtpEnabled || vlmMtpEnabled || dflashEnabled || specprefillEnabled
-            || turboquantKvEnabled || qwen35AnePrefillEnabled || thinkingBudgetEnabled
-            || model?.settings?.guidedGrammarEnabled == true
-            || penalties.contains(where: { !$0.0.isEmpty && Double($0.0) != $0.1 }) {
-            return String(localized: "settings.uno.conflict",
-                          defaultValue: "Disable other decode accelerators, grammar constraints and thinking budgets. Clear Min P and penalty settings.")
-        }
-        return nil
-    }
-
-    var unoUnavailableReason: String? {
-        unoConflictReason ?? (unoAdapterCandidates.isEmpty
-            ? String(localized: "settings.uno.missing", defaultValue: "No compatible Uno adapter is available.")
-            : nil)
-    }
-
-    func validateUnoWorkingSettings() -> Bool {
-        guard unoEnabled else { return true }
-        if let reason = unoConflictReason {
-            lastError = reason
-            return false
-        }
-        guard unoAdapterCandidates.contains(where: { $0.id == unoAdapterModel }) else {
-            lastError = String(localized: "settings.uno.select", defaultValue: "Select a Uno adapter")
-            return false
-        }
-        return true
-    }
-
     var isQwen4Exp: Bool {
         (model?.configModelType ?? "")
             .lowercased()
@@ -495,7 +450,7 @@ final class ModelSettingsScreenVM {
             return true
         case .dflashSsdCache, .dflashSsdCacheGib:
             return true
-        case .mtpEnabled, .unoEnabled, .unoAdapterModel, .vlmMtpEnabled, .vlmMtpDraftModel:
+        case .mtpEnabled, .vlmMtpEnabled, .vlmMtpDraftModel:
             return true
         case .vlmMtpDraftBlockSize:
             return true
@@ -651,8 +606,6 @@ final class ModelSettingsScreenVM {
                 self.dflashSsdCache = s?.dflashSsdCache ?? false
                 self.dflashSsdCacheGib = DflashByteSize.bytesToGib(s?.dflashSsdCacheMaxBytes)
                     .map(String.init) ?? "20"
-                self.unoEnabled = s?.unoEnabled ?? false
-                self.unoAdapterModel = s?.unoAdapterModel ?? ""
                 self.mtpEnabled = s?.mtpEnabled ?? false
                 self.vlmMtpEnabled = s?.vlmMtpEnabled ?? false
                 self.vlmMtpDraftModel = s?.vlmMtpDraftModel ?? ""
@@ -889,10 +842,6 @@ final class ModelSettingsScreenVM {
         case .dflashSsdCacheGib:
             patch.dflashSsdCacheMaxBytes = DflashByteSize.gibToBytes(Int(dflashSsdCacheGib))
         case .mtpEnabled:              patch.mtpEnabled = mtpEnabled
-        case .unoEnabled, .unoAdapterModel:
-            guard isK2Base, validateUnoWorkingSettings() else { return }
-            patch.unoEnabled = unoEnabled
-            patch.unoAdapterModel = unoAdapterModel
         case .vlmMtpEnabled:           patch.vlmMtpEnabled = vlmMtpEnabled
         case .vlmMtpDraftModel:        patch.vlmMtpDraftModel = vlmMtpDraftModel.isEmpty ? nil : vlmMtpDraftModel
         case .vlmMtpDraftBlockSize:    patch.vlmMtpDraftBlockSize = Int(vlmMtpDraftBlockSize)
@@ -1320,10 +1269,6 @@ final class ModelSettingsScreenVM {
                     out[ProfileSettingsKey.dflashSsdCacheMaxBytes] = AnyCodable(Int(bytes))
                 }
             }
-            if isK2Base {
-                putBool(ProfileSettingsKey.unoEnabled, unoEnabled)
-                putString(ProfileSettingsKey.unoAdapterModel, unoAdapterModel)
-            }
             putBool(ProfileSettingsKey.mtpEnabled, mtpEnabled)
             putBool(ProfileSettingsKey.vlmMtpEnabled, vlmMtpEnabled)
             if vlmMtpEnabled {
@@ -1461,7 +1406,7 @@ final class ModelSettingsScreenVM {
     func saveWorkingAs(scope: ProfileScope, name: String, client: OMLXClient) async {
         let cleanName = name.trimmingCharacters(in: .whitespaces)
         guard !cleanName.isEmpty, scope != .preset else { return }
-        guard validateUnoWorkingSettings(), validateK2AneWorkingSettings(), validateQwenAneWorkingSettings() else { return }
+        guard validateK2AneWorkingSettings(), validateQwenAneWorkingSettings() else { return }
         let settings = currentSettingsDict()
         do {
             switch scope {
@@ -1508,7 +1453,7 @@ final class ModelSettingsScreenVM {
     /// ProfileDetailCard preview's "Update with working" button.
     func updateProfileWithWorking(scope: ProfileScope, name: String, client: OMLXClient) async {
         guard scope != .preset else { return }
-        guard validateUnoWorkingSettings(), validateK2AneWorkingSettings(), validateQwenAneWorkingSettings() else { return }
+        guard validateK2AneWorkingSettings(), validateQwenAneWorkingSettings() else { return }
         let settings = currentSettingsDict()
         do {
             switch scope {
