+++ apps/omlx-mac/Tests/oMLXTests/LocalizationSmokeTests.swift
@@ -73,7 +73,6 @@ final class LocalizationSmokeTests: XCTestCase {
         "bench.context.header.title", "bench.context.section.configuration",
         // Settings + helpers
         "settings.section.basic", "settings.advanced.experimental.section",
-        "settings.uno.enable", "settings.uno.adapter", "settings.uno.conflict",
         "appearance.row.menubar_icon", "appearance.row.menubar_icon.restore",
         // Menubar + updates
         "menubar.item.quit", "menubar.stats.session_section",
