+++ apps/omlx-mac/Tests/oMLXTests/MenubarControllerModelsTests.swift
@@ -215,7 +215,6 @@ final class MenubarControllerModelsTests: XCTestCase {
             engineType: nil,
             modelType: nil,
             configModelType: nil,
-            unoBaseModelId: nil,
             modelContextLength: nil,
             thinkingDefault: nil,
             dflashCompatible: nil,
