+++ omlx/admin/ane_tuning.py
@@ -430,7 +430,6 @@ def _settings_for_candidate(base: Any, request: ANETuningRequest, candidate: _Ca
         settings.k2_ane_prefill_fraction = candidate.mlp_fraction or 1 / 3
         settings.k2_ane_prefill_shared_fraction = candidate.shared_fraction
         for field in (
-            "uno_enabled",
             "qwen35_ane_prefill_enabled",
             "dflash_enabled",
             "specprefill_enabled",
