+++ omlx/admin/i18n/en.json
@@ -608,7 +608,6 @@
   "modal.model_settings.loading_defaults": "Loading...",
   "modal.model_settings.placeholder_default": "Default",
   "modal.model_settings.k2_thinking_hint": "K2 uses reasoning effort. Use Thinking Budget to limit reasoning.",
-  "modal.model_settings.uno_budget_hint": "Uno does not support thinking budgets.",
   "modal.model_settings.enable_thinking": "Enable Thinking",
   "modal.model_settings.enable_thinking_hint": "Enable reasoning/thinking mode for this model.",
   "modal.model_settings.enable_thinking_default": "Using model default",
@@ -679,12 +678,6 @@
   "modal.model_settings.mtplx_import": "Import MTP side-car",
   "modal.model_settings.mtplx_importing": "Importing...",
   "modal.model_settings.mtp_conflict_turboquant": "Disable TurboQuant KV first.",
-  "modal.model_settings.uno": "Enable Uno",
-  "modal.model_settings.uno_hint": "Use a conditional diffusion adapter with this base model.",
-  "modal.model_settings.uno_missing": "No compatible local Uno adapter is available.",
-  "modal.model_settings.uno_conflict": "Disable other decode accelerators, grammar constraints and thinking budgets. Clear Min P and penalty settings.",
-  "modal.model_settings.uno_adapter": "Uno adapter",
-  "modal.model_settings.uno_select": "Select an adapter",
   "modal.model_settings.vlm_mtp": "VLM MTP",
   "modal.model_settings.vlm_mtp_hint": "Speculative decoding via an external MTP drafter model (Gemma 4 assistant or Qwen MTP). ~1.3–1.6× single-request speedup; concurrent eligible requests fall back to standard batching.",
   "modal.model_settings.vlm_mtp_conflict": "Disable DFlash / SpecPrefill / MTP / TurboQuant KV first — VLM MTP owns the speculative decode path.",
