+++ omlx/admin/i18n/es.json
@@ -608,7 +608,6 @@
   "modal.model_settings.loading_defaults": "Cargando...",
   "modal.model_settings.placeholder_default": "Predeterminado",
   "modal.model_settings.k2_thinking_hint": "K2 usa el esfuerzo de razonamiento. Use Presupuesto de Pensamiento para limitar el razonamiento.",
-  "modal.model_settings.uno_budget_hint": "Uno no admite presupuestos de pensamiento.",
   "modal.model_settings.enable_thinking": "Habilitar Pensamiento",
   "modal.model_settings.enable_thinking_hint": "Habilitar modo de razonamiento/pensamiento para este modelo.",
   "modal.model_settings.enable_thinking_default": "Usando el predeterminado del modelo",
@@ -679,12 +678,6 @@
   "modal.model_settings.mtplx_import": "Importar side-car MTP",
   "modal.model_settings.mtplx_importing": "Importando...",
   "modal.model_settings.mtp_conflict_turboquant": "Disable TurboQuant KV first.",
-  "modal.model_settings.uno": "Habilitar Uno",
-  "modal.model_settings.uno_hint": "Use un adaptador de difusión condicional con este modelo base.",
-  "modal.model_settings.uno_missing": "No hay ningún adaptador Uno local compatible disponible.",
-  "modal.model_settings.uno_conflict": "Desactive los demás aceleradores de decodificación, las restricciones gramaticales y los presupuestos de pensamiento. Borre los ajustes de Min P y penalizaciones.",
-  "modal.model_settings.uno_adapter": "Adaptador Uno",
-  "modal.model_settings.uno_select": "Seleccione un adaptador",
   "modal.model_settings.vlm_mtp": "VLM MTP",
   "modal.model_settings.vlm_mtp_hint": "Speculative decoding via an external MTP drafter model (Gemma 4 assistant or Qwen MTP). ~1.3–1.6× single-request speedup; concurrent eligible requests fall back to standard batching.",
   "modal.model_settings.vlm_mtp_conflict": "Disable DFlash / SpecPrefill / MTP / TurboQuant KV first — VLM MTP owns the speculative decode path.",
