+++ omlx/admin/i18n/fr.json
@@ -608,7 +608,6 @@
   "modal.model_settings.loading_defaults": "Chargement...",
   "modal.model_settings.placeholder_default": "Par défaut",
   "modal.model_settings.k2_thinking_hint": "K2 utilise un effort de raisonnement. Utilisez Budget de réflexion pour limiter le raisonnement.",
-  "modal.model_settings.uno_budget_hint": "Uno ne prend pas en charge les budgets de réflexion.",
   "modal.model_settings.enable_thinking": "Activer la réflexion",
   "modal.model_settings.enable_thinking_hint": "Activez le mode raisonnement/réflexion pour ce modèle.",
   "modal.model_settings.enable_thinking_default": "Utilisant la valeur par défaut du modèle",
@@ -679,12 +678,6 @@
   "modal.model_settings.mtplx_import": "Importer le side-car MTP",
   "modal.model_settings.mtplx_importing": "Importation...",
   "modal.model_settings.mtp_conflict_turboquant": "Disable TurboQuant KV first.",
-  "modal.model_settings.uno": "Activer Uno",
-  "modal.model_settings.uno_hint": "Utilisez un adaptateur de diffusion conditionnelle avec ce modèle de base.",
-  "modal.model_settings.uno_missing": "Aucun adaptateur Uno local compatible n’est disponible.",
-  "modal.model_settings.uno_conflict": "Désactivez les autres accélérateurs de décodage, les contraintes grammaticales et les budgets de réflexion. Effacez les paramètres Min P et les pénalités.",
-  "modal.model_settings.uno_adapter": "Adaptateur Uno",
-  "modal.model_settings.uno_select": "Sélectionnez un adaptateur",
   "modal.model_settings.vlm_mtp": "VLM MTP",
   "modal.model_settings.vlm_mtp_hint": "Speculative decoding via an external MTP drafter model (Gemma 4 assistant or Qwen MTP). ~1.3–1.6× single-request speedup; concurrent eligible requests fall back to standard batching.",
   "modal.model_settings.vlm_mtp_conflict": "Disable DFlash / SpecPrefill / MTP / TurboQuant KV first — VLM MTP owns the speculative decode path.",
