+++ omlx/admin/i18n/ja.json
@@ -608,7 +608,6 @@
   "modal.model_settings.loading_defaults": "読み込み中...",
   "modal.model_settings.placeholder_default": "デフォルト",
   "modal.model_settings.k2_thinking_hint": "K2 は推論の強度を使用します。「思考トークン制限」で推論を制限できます。",
-  "modal.model_settings.uno_budget_hint": "Uno は思考トークン制限に対応していません。",
   "modal.model_settings.enable_thinking": "思考モード",
   "modal.model_settings.enable_thinking_hint": "このモデルの推論/思考モードを設定します。",
   "modal.model_settings.enable_thinking_default": "モデルのデフォルトを使用中",
@@ -679,12 +678,6 @@
   "modal.model_settings.mtplx_import": "MTPサイドカーをインポート",
   "modal.model_settings.mtplx_importing": "インポート中...",
   "modal.model_settings.mtp_conflict_turboquant": "Disable TurboQuant KV first.",
-  "modal.model_settings.uno": "Uno を有効にする",
-  "modal.model_settings.uno_hint": "このベースモデルで条件付き拡散アダプターを使用します。",
-  "modal.model_settings.uno_missing": "互換性のあるローカルの Uno アダプターがありません。",
-  "modal.model_settings.uno_conflict": "他のデコード高速化機能、文法制約、思考トークン制限を無効にしてください。Min P とペナルティの設定をクリアしてください。",
-  "modal.model_settings.uno_adapter": "Uno アダプター",
-  "modal.model_settings.uno_select": "アダプターを選択",
   "modal.model_settings.vlm_mtp": "VLM MTP",
   "modal.model_settings.vlm_mtp_hint": "Speculative decoding via an external MTP drafter model (Gemma 4 assistant or Qwen MTP). ~1.3–1.6× single-request speedup; concurrent eligible requests fall back to standard batching.",
   "modal.model_settings.vlm_mtp_conflict": "Disable DFlash / SpecPrefill / MTP / TurboQuant KV first — VLM MTP owns the speculative decode path.",
