+++ omlx/admin/i18n/ko.json
@@ -608,7 +608,6 @@
   "modal.model_settings.loading_defaults": "로딩 중...",
   "modal.model_settings.placeholder_default": "기본값",
   "modal.model_settings.k2_thinking_hint": "K2는 추론 강도를 사용합니다. Thinking 토큰 제한으로 추론을 제한하세요.",
-  "modal.model_settings.uno_budget_hint": "Uno는 Thinking 토큰 제한을 지원하지 않습니다.",
   "modal.model_settings.enable_thinking": "Thinking 모드",
   "modal.model_settings.enable_thinking_hint": "이 모델의 추론/thinking 모드를 설정합니다.",
   "modal.model_settings.enable_thinking_default": "모델 기본값 사용 중",
@@ -679,12 +678,6 @@
   "modal.model_settings.mtplx_import": "MTP 사이드카 가져오기",
   "modal.model_settings.mtplx_importing": "가져오는 중...",
   "modal.model_settings.mtp_conflict_turboquant": "TurboQuant KV 를 먼저 비활성화하세요.",
-  "modal.model_settings.uno": "Uno 활성화",
-  "modal.model_settings.uno_hint": "이 기본 모델에 조건부 확산 어댑터를 사용합니다.",
-  "modal.model_settings.uno_missing": "호환되는 로컬 Uno 어댑터가 없습니다.",
-  "modal.model_settings.uno_conflict": "다른 디코딩 가속 기능, 문법 제약 및 Thinking 토큰 제한을 비활성화하세요. Min P와 페널티 설정을 지우세요.",
-  "modal.model_settings.uno_adapter": "Uno 어댑터",
-  "modal.model_settings.uno_select": "어댑터 선택",
   "modal.model_settings.vlm_mtp": "VLM MTP",
   "modal.model_settings.vlm_mtp_hint": "Speculative decoding via an external MTP drafter model (Gemma 4 assistant or Qwen MTP). ~1.3–1.6× single-request speedup; concurrent eligible requests fall back to standard batching.",
   "modal.model_settings.vlm_mtp_conflict": "Disable DFlash / SpecPrefill / MTP / TurboQuant KV first — VLM MTP owns the speculative decode path.",
