+++ omlx/admin/i18n/pt-BR.json
@@ -608,7 +608,6 @@
   "modal.model_settings.loading_defaults": "Carregando...",
   "modal.model_settings.placeholder_default": "Padrão",
   "modal.model_settings.k2_thinking_hint": "O K2 usa o esforço de raciocínio. Use Orçamento de Raciocínio para limitar o raciocínio.",
-  "modal.model_settings.uno_budget_hint": "O Uno não oferece suporte a orçamentos de raciocínio.",
   "modal.model_settings.enable_thinking": "Ativar Raciocínio",
   "modal.model_settings.enable_thinking_hint": "Ativar modo de raciocínio/thinking para este modelo.",
   "modal.model_settings.enable_thinking_default": "Usando padrão do modelo",
@@ -679,12 +678,6 @@
   "modal.model_settings.mtplx_import": "Importar side-car MTP",
   "modal.model_settings.mtplx_importing": "Importando...",
   "modal.model_settings.mtp_conflict_turboquant": "Desative o TurboQuant KV primeiro.",
-  "modal.model_settings.uno": "Ativar Uno",
-  "modal.model_settings.uno_hint": "Use um adaptador de difusão condicional com este modelo base.",
-  "modal.model_settings.uno_missing": "Nenhum adaptador Uno local compatível está disponível.",
-  "modal.model_settings.uno_conflict": "Desative os outros aceleradores de decodificação, as restrições gramaticais e os orçamentos de raciocínio. Limpe as configurações de Min P e penalidades.",
-  "modal.model_settings.uno_adapter": "Adaptador Uno",
-  "modal.model_settings.uno_select": "Selecione um adaptador",
   "modal.model_settings.vlm_mtp": "VLM MTP",
   "modal.model_settings.vlm_mtp_hint": "Decodificação especulativa via modelo rascunhador MTP externo (assistente Gemma 4 ou MTP Qwen). ~1,3–1,6× mais rápido para requisições únicas; requisições concorrentes elegíveis fazem fallback para o batching padrão.",
   "modal.model_settings.vlm_mtp_conflict": "Desative DFlash / SpecPrefill / MTP / TurboQuant KV primeiro — o VLM MTP assume o caminho de decodificação especulativa.",
