+++ omlx/admin/i18n/ru.json
@@ -608,7 +608,6 @@
   "modal.model_settings.loading_defaults": "Загрузка...",
   "modal.model_settings.placeholder_default": "По умолчанию",
   "modal.model_settings.k2_thinking_hint": "K2 использует уровень рассуждения. Используйте «Бюджет мышления», чтобы ограничить рассуждение.",
-  "modal.model_settings.uno_budget_hint": "Uno не поддерживает бюджет мышления.",
   "modal.model_settings.enable_thinking": "Включить мышление",
   "modal.model_settings.enable_thinking_hint": "Включить режим рассуждения для этой модели.",
   "modal.model_settings.enable_thinking_default": "Использовать значение модели по умолчанию",
@@ -679,12 +678,6 @@
   "modal.model_settings.mtplx_import": "Импортировать side-car MTP",
   "modal.model_settings.mtplx_importing": "Импорт...",
   "modal.model_settings.mtp_conflict_turboquant": "Сначала отключите TurboQuant KV.",
-  "modal.model_settings.uno": "Включить Uno",
-  "modal.model_settings.uno_hint": "Используйте адаптер условной диффузии с этой базовой моделью.",
-  "modal.model_settings.uno_missing": "Нет совместимого локального адаптера Uno.",
-  "modal.model_settings.uno_conflict": "Отключите другие ускорители декодирования, грамматические ограничения и бюджет мышления. Очистите настройки Min P и штрафов.",
-  "modal.model_settings.uno_adapter": "Адаптер Uno",
-  "modal.model_settings.uno_select": "Выберите адаптер",
   "modal.model_settings.vlm_mtp": "VLM MTP",
   "modal.model_settings.vlm_mtp_hint": "Спекулятивное декодирование через внешнюю MTP-модель-драфтер (ассистент Gemma 4 или Qwen MTP). Ускоряет одиночные запросы примерно в 1.3–1.6×; параллельные подходящие запросы переходят на стандартную пакетную обработку.",
   "modal.model_settings.vlm_mtp_conflict": "Сначала отключите DFlash / SpecPrefill / MTP / TurboQuant KV: VLM MTP управляет путем спекулятивного декодирования.",
