+++ omlx/admin/i18n/zh-TW.json
@@ -608,7 +608,6 @@
   "modal.model_settings.loading_defaults": "載入中...",
   "modal.model_settings.placeholder_default": "預設",
   "modal.model_settings.k2_thinking_hint": "K2 使用推理強度。可透過「思考 Token 限制」限制推理。",
-  "modal.model_settings.uno_budget_hint": "Uno 不支援思考 Token 限制。",
   "modal.model_settings.enable_thinking": "啟用思考",
   "modal.model_settings.enable_thinking_hint": "設定此模型的推理/思考模式。",
   "modal.model_settings.enable_thinking_default": "使用模型預設值",
@@ -679,12 +678,6 @@
   "modal.model_settings.mtplx_import": "匯入 MTP 側掛權重",
   "modal.model_settings.mtplx_importing": "匯入中...",
   "modal.model_settings.mtp_conflict_turboquant": "請先停用 TurboQuant KV。",
-  "modal.model_settings.uno": "啟用 Uno",
-  "modal.model_settings.uno_hint": "為此基礎模型使用條件擴散適配器。",
-  "modal.model_settings.uno_missing": "沒有相容的本機 Uno 適配器。",
-  "modal.model_settings.uno_conflict": "請停用其他解碼加速功能、文法限制和思考 Token 限制，並清除 Min P 和懲罰設定。",
-  "modal.model_settings.uno_adapter": "Uno 適配器",
-  "modal.model_settings.uno_select": "選擇適配器",
   "modal.model_settings.vlm_mtp": "VLM MTP",
   "modal.model_settings.vlm_mtp_hint": "透過外部 MTP drafter 模型進行推測解碼（Gemma 4 assistant 或 Qwen MTP）。單一請求約 1.3–1.6× 加速；並發符合條件的請求退回至標準批次處理。",
   "modal.model_settings.vlm_mtp_conflict": "請先停用 DFlash / SpecPrefill / MTP / TurboQuant KV — VLM MTP 獨佔推測解碼路徑。",
