+++ omlx/admin/i18n/zh.json
@@ -608,7 +608,6 @@
   "modal.model_settings.loading_defaults": "加载中...",
   "modal.model_settings.placeholder_default": "默认",
   "modal.model_settings.k2_thinking_hint": "K2 使用推理强度。可通过“思考 Token 限制”限制推理。",
-  "modal.model_settings.uno_budget_hint": "Uno 不支持思考 Token 限制。",
   "modal.model_settings.enable_thinking": "启用思考",
   "modal.model_settings.enable_thinking_hint": "设置此模型的推理/思考模式。",
   "modal.model_settings.enable_thinking_default": "使用模型默认值",
@@ -679,12 +678,6 @@
   "modal.model_settings.mtplx_import": "导入 MTP 边车权重",
   "modal.model_settings.mtplx_importing": "正在导入...",
   "modal.model_settings.mtp_conflict_turboquant": "请先禁用 TurboQuant KV。",
-  "modal.model_settings.uno": "启用 Uno",
-  "modal.model_settings.uno_hint": "为此基础模型使用条件扩散适配器。",
-  "modal.model_settings.uno_missing": "没有兼容的本地 Uno 适配器。",
-  "modal.model_settings.uno_conflict": "请禁用其他解码加速功能、语法约束和思考 Token 限制，并清除 Min P 和惩罚设置。",
-  "modal.model_settings.uno_adapter": "Uno 适配器",
-  "modal.model_settings.uno_select": "选择适配器",
   "modal.model_settings.vlm_mtp": "VLM MTP",
   "modal.model_settings.vlm_mtp_hint": "通过外部 MTP drafter 模型进行推测解码（支持 Gemma 4 assistant 和 Qwen MTP）。单请求约 1.3–1.6× 加速；并发请求回退到标准批处理。",
   "modal.model_settings.vlm_mtp_conflict": "请先禁用 DFlash / SpecPrefill / MTP / TurboQuant KV — VLM MTP 独占推测解码路径。",
