+++ omlx/admin/templates/dashboard/_modal_model_settings.html
@@ -481,9 +481,9 @@
                                 <div class="flex items-start justify-between gap-3">
                                     <div class="min-w-0">
                                         <span class="text-sm font-medium text-neutral-700">{{ t('modal.model_settings.thinking_budget') }}</span>
-                                        <p class="text-xs text-neutral-500 mt-0.5" x-text="modelSettings.uno_enabled ? window.t('modal.model_settings.uno_budget_hint') : window.t('modal.model_settings.thinking_budget_hint')"></p>
+                                        <p class="text-xs text-neutral-500 mt-0.5" x-text="window.t('modal.model_settings.thinking_budget_hint')"></p>
                                     </div>
-                                    <button type="button" :disabled="modelSettings.uno_enabled && !modelSettings.enableThinkingBudget" @click="modelSettings.enableThinkingBudget = !modelSettings.enableThinkingBudget; if (!modelSettings.enableThinkingBudget) modelSettings.thinking_budget_tokens = null;"
+                                    <button type="button" @click="modelSettings.enableThinkingBudget = !modelSettings.enableThinkingBudget; if (!modelSettings.enableThinkingBudget) modelSettings.thinking_budget_tokens = null;"
                                             :class="modelSettings.enableThinkingBudget ? 'bg-black' : 'bg-neutral-200'"
                                             class="relative flex-shrink-0 w-11 h-6 mt-0.5 rounded-full transition-colors duration-300 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-black">
                                         <span :class="modelSettings.enableThinkingBudget ? 'translate-x-5' : 'translate-x-0'"
@@ -491,7 +491,7 @@
                                     </button>
                                 </div>
                                 <div x-show="modelSettings.enableThinkingBudget" x-transition class="pt-1">
-                                    <input type="number" :disabled="modelSettings.uno_enabled" x-model.number="modelSettings.thinking_budget_tokens" placeholder="{{ t('modal.model_settings.thinking_budget_placeholder') }}" min="1" step="100"
+                                    <input type="number" x-model.number="modelSettings.thinking_budget_tokens" placeholder="{{ t('modal.model_settings.thinking_budget_placeholder') }}" min="1" step="100"
                                            class="w-full px-4 py-2.5 border border-neutral-200 rounded-xl text-sm focus:ring-2 focus:ring-neutral-900 focus:border-transparent transition-all">
                                 </div>
                             </div>
@@ -1343,37 +1343,6 @@
                                                    class="w-full px-4 py-2.5 border border-neutral-200 rounded-xl text-sm focus:ring-2 focus:ring-neutral-900 focus:border-transparent transition-all">
                                             <p class="text-xs text-neutral-400 mt-1">Disk budget for L2 spill; oldest entries are evicted when exceeded.</p>
                                         </div>
-                                    </div>
-                                </div>
-
-                                <div x-show="selectedModel?.config_model_type === 'k2_horizon' && !selectedModel?.is_helper"
-                                     class="p-4 bg-neutral-50 rounded-xl space-y-3 mt-3">
-                                    <div class="flex items-start justify-between gap-3">
-                                        <div>
-                                            <span class="text-sm font-medium text-neutral-700">{{ t('modal.model_settings.uno') }}</span>
-                                            <p class="text-xs text-neutral-500 mt-0.5">{{ t('modal.model_settings.uno_hint') }}</p>
-                                            <p x-show="!unoAdapterCandidates().length" class="text-xs text-amber-600 mt-1">{{ t('modal.model_settings.uno_missing') }}</p>
-                                            <p x-show="unoConflict()" class="text-xs text-amber-600 mt-1">{{ t('modal.model_settings.uno_conflict') }}</p>
-                                        </div>
-                                        <button type="button" role="switch" :aria-checked="!!modelSettings.uno_enabled"
-                                                aria-label="{{ t('modal.model_settings.uno') }}"
-                                                @click="modelSettings.uno_enabled = !modelSettings.uno_enabled"
-                                                :disabled="!modelSettings.uno_enabled && (!unoAdapterCandidates().length || unoConflict())"
-                                                :class="modelSettings.uno_enabled ? 'bg-black' : 'bg-neutral-200'"
-                                                class="relative flex-shrink-0 w-11 h-6 rounded-full disabled:opacity-40 focus:outline-none focus:ring-2 focus:ring-black">
-                                            <span :class="modelSettings.uno_enabled ? 'translate-x-5' : 'translate-x-0'"
-                                                  class="block w-5 h-5 bg-white rounded-full shadow-sm transform transition-transform absolute top-0.5 left-0.5"></span>
-                                        </button>
-                                    </div>
-                                    <div x-show="modelSettings.uno_enabled">
-                                        <label for="uno-adapter" class="block text-xs font-medium text-neutral-600">{{ t('modal.model_settings.uno_adapter') }}</label>
-                                        <select id="uno-adapter" x-model="modelSettings.uno_adapter_model"
-                                                class="w-full px-3 py-2 text-sm border border-neutral-200 rounded-lg focus:ring-2 focus:ring-black bg-white">
-                                            <option value="">{{ t('modal.model_settings.uno_select') }}</option>
-                                            <template x-for="m in unoAdapterCandidates()" :key="m.id">
-                                                <option :value="m.id" x-text="m.id"></option>
-                                            </template>
-                                        </select>
                                     </div>
                                 </div>
