+++ omlx/engine_pool.py
@@ -375,11 +375,6 @@ class EnginePool:
                 format_size(extra),
                 entry.model_id,
             )
-        if getattr(runtime_settings, "uno_enabled", False):
-            adapter = self._entries.get(runtime_settings.uno_adapter_model)
-            if adapter is None or adapter.config_model_type != "k2_horizon_uno":
-                raise ValueError("Select an available Uno adapter.")
-            base = entry.estimated_size + adapter.estimated_size
         if getattr(runtime_settings, "k2_ane_prefill_enabled", False):
             from .patches.k2_horizon.ane_prefill import prefill_memory_reservation
 
@@ -774,9 +769,6 @@ class EnginePool:
             add("vlm_mtp_draft_model", data.get("vlm_mtp_draft_model"))
             add("vlm_mtp_draft_block_size", data.get("vlm_mtp_draft_block_size"))
 
-        add("uno_enabled", bool(data.get("uno_enabled", False)))
-        if data.get("uno_enabled"):
-            add("uno_adapter_model", data.get("uno_adapter_model"))
         add("k2_ane_prefill_enabled", bool(data.get("k2_ane_prefill_enabled", False)))
         if data.get("k2_ane_prefill_enabled"):
             for key, default in (
@@ -2652,8 +2644,6 @@ class EnginePool:
         pre_load_memory = max(mx.get_active_memory(), get_phys_footprint())
         try:
             effective_type = entry.engine_type
-            if entry.config_model_type == "k2_horizon_uno":
-                raise ValueError("Select a compatible K2 base and enable Uno in its settings.")
             if force_lm and effective_type == "vlm":
                 effective_type = "batched"
                 logger.info(f"Loading model as LM (force_lm=True): {model_id}")
@@ -2667,9 +2657,6 @@ class EnginePool:
             model_settings = self._effective_qwen4_model_settings(entry, model_settings)
 
             deployment = self._distributed_deployment_for_entry(entry)
-            uno_enabled = bool(getattr(model_settings, "uno_enabled", False))
-            if uno_enabled and deployment is not None:
-                raise ValueError("Uno does not support distributed deployment.")
             base_resident_size = self._entry_resident_size(entry)
             if (
                 deployment is None
@@ -2705,11 +2692,7 @@ class EnginePool:
             # since DFlash has its own model loading pipeline
             engine = None
             deployment = deployment if effective_type == "batched" else None
-            if (
-                not uno_enabled
-                and deployment is None
-                and model_settings is not None
-            ):
+            if deployment is None and model_settings is not None:
                 dflash_enabled = getattr(model_settings, "dflash_enabled", False)
                 dflash_draft = getattr(model_settings, "dflash_draft_model", None)
                 if (
@@ -2784,17 +2767,7 @@ class EnginePool:
 
             # Create engine based on engine type (if DFlash not active)
             if engine is None:
-                if uno_enabled:
-                    from .engine.uno import UnoEngine
-
-                    adapter = self._entries[model_settings.uno_adapter_model]
-                    engine = UnoEngine(
-                        model_name=entry.model_path,
-                        adapter_path=adapter.model_path,
-                        scheduler_config=self._scheduler_config,
-                        model_settings=model_settings,
-                    )
-                elif deployment is not None:
+                if deployment is not None:
                     from .engine.distributed import DistributedBatchedEngine
 
                     deployment = replace(
