+++ omlx/patches/k2_horizon/tool_grammar.py
@@ -90,44 +90,3 @@ def compile_tool_grammar(compiler, tools, existing=None):
             },
         }
     )
-
-
-class UnoToolConstraint:
-    """Advance speculative matchers independently and commit only accepted tokens."""
-
-    def __init__(self, compiled_grammar, vocab_size):
-        import numpy as np
-        import xgrammar as xgr
-        from xgrammar.kernels.apply_token_bitmask_mlx import apply_token_bitmask_mlx
-
-        self.matcher = xgr.GrammarMatcher(compiled_grammar)
-        self.bitmask = np.full((1, (vocab_size + 31) // 32), -1, dtype=np.int32)
-        self.apply_mask = apply_token_bitmask_mlx
-        self.vocab_size = vocab_size
-
-    def seed(self, logits):
-        import mlx.core as mx
-
-        self.bitmask.fill(-1)
-        self.matcher.fill_next_token_bitmask(self.bitmask)
-        first = self.apply_mask(mx.array(self.bitmask), logits[:1], self.vocab_size)
-        return mx.concatenate([first, logits[1:]])
-
-    def verify(self, logits, proposals):
-        import mlx.core as mx
-        import numpy as np
-
-        matcher = self.matcher.fork()
-        vocab_size = logits.shape[-1]
-        masks = np.full((len(proposals), (vocab_size + 31) // 32), -1, dtype=np.int32)
-        for row, token in enumerate(proposals):
-            if matcher.is_terminated() or not matcher.accept_token(token):
-                break
-            if not matcher.is_terminated():
-                matcher.fill_next_token_bitmask(masks, row)
-        return self.apply_mask(mx.array(masks), logits, vocab_size)
-
-    def commit(self, tokens):
-        for token in tokens:
-            if not self.matcher.accept_token(token):
-                raise RuntimeError("Uno committed a token outside the K2 tool grammar")
