+++ tests/test_k2_ane_ui_presets.py
@@ -15,17 +15,26 @@ def test_fraction_choices_preserve_current_value(shared, current):
     node = shutil.which("node")
     if not node:
         pytest.skip("Node.js is required to exercise dashboard JavaScript")
-    source = (Path(__file__).parents[1] / "omlx/admin/static/js/dashboard.js").read_text()
+    source = (
+        Path(__file__).parents[1] / "omlx/admin/static/js/dashboard.js"
+    ).read_text()
     method = re.search(
         r"^            k2AneFractionOptions\([^\n]*\) \{.*?^            \},",
-        source, re.M | re.S,
+        source,
+        re.M | re.S,
     ).group()
-    script = ("const app = {" + method + "}; console.log(JSON.stringify("
-              + f"app.k2AneFractionOptions({json.dumps(current)}, {json.dumps(shared)})"
-              + "));")
+    script = (
+        "const app = {"
+        + method
+        + "}; console.log(JSON.stringify("
+        + f"app.k2AneFractionOptions({json.dumps(current)}, {json.dumps(shared)})"
+        + "));"
+    )
     options = json.loads(subprocess.check_output([node, "-e", script], text=True))
     expected = [0, 1 / 3, 1] if shared else [1 / 3, 0.5]
     if current not in expected:
         expected.insert(0, current)
     assert [option["value"] for option in options] == expected
-    assert next(option["value"] for option in options if option["label"] == "33%") == 1 / 3
+    assert (
+        next(option["value"] for option in options if option["label"] == "33%") == 1 / 3
+    )
